% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_chalkboard.R
\name{theme_chalkboard}
\alias{theme_chalkboard}
\title{Title}
\usage{
theme_chalkboard(
  board_color = "darkseagreen4",
  chalk_color = "lightyellow",
  ...
)
}
\arguments{
\item{board_color}{a string which is a color value}

\item{chalk_color}{a string which is a color value}

\item{...}{}
}
\value{
a ggplot2 object
}
\description{
Title
}
