% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_chalkboard.R
\name{theme_chalkboard_slate}
\alias{theme_chalkboard_slate}
\title{Title}
\usage{
theme_chalkboard_slate()
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
geoms_chalk()
ggplot(cars) +
aes(speed, dist) +
geom_point() +
theme_chalkboard_slate()
}
