% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_chalkboard.R
\name{theme_chalkboard}
\alias{theme_chalkboard}
\title{Title}
\usage{
theme_chalkboard(board_color = "darkseagreen4", chalk_color = "lightyellow")
}
\arguments{
\item{chalk_color}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
geoms_chalk()
ggplot(cars) +
aes(speed, dist) +
geom_point() +
theme_chalkboard()
}
